Move traverseFragmentInstanceChildren to internal ReactFiberTreeReflection
